Fix assertoin failure for certain kinds of merges
The RoundTripConstructor.flatten_mapping() method made an assumption that
when merging a mapping into another mapping, the actual data would have been
defined already. For example:

  predefined: &contents { a: 1 }
  merged:
    <<: *contents

However, the following is valid as well, and in this case the data is only
defined at the point of merging:

  merged-1:
    <<: &contents  { a: 1 }
  merged-2:
    <<: *contents

This commit fixes an assertion failure for the latter case.